policy/ip_check: remove uniqueItems attr from schema
This is just because of how the react-jsonschema-form library renders an
array with uniqueItems=true. It turns out that it renders a multiple
selection box, but it does not allow to sort its elements.

The original intention of the policy was to allow sorting those elements
to try them in the selected order. Deleting uniqueItems=true solves the
issue. However, now it's possible to add duplicates in the array but it
does not really affect the behavior of the policy.