[THREESCALE-2264] Obtain JWT client_id from different claim
This commit adds the option to obtain the client_id from a different claim and
does not use the AZP claim if the identity provider does not need that. This
commit adds a new option `oidc_client_id` that can be used as a liquid template
to retrieve any claim from the JWT payload.

Fixes #1008
Fixes THREESCALE-2264

Signed-off-by: Eloy Coto <eloy.coto@gmail.com>