[Test] Luassert make sure to revert function if it is already mocked.
This commit overwrites luassert stub function to make sure that before
try to mock any function verifies that is not mocked, and if it is,
revert to the original one. This prevents scenarios where multiple stubs
are in one test and are leaked to other test and that scenario is hard
to debug/notice.

Signed-off-by: Eloy Coto <eloy.coto@gmail.com>