Policy: Rate-limit fix window exception
On rate-limit window can be set as 0 so the division that happens on
resty will fail due to division by 0 is not allowed.

This commit fixes that and validates that window value is bigger than 0,
if not set the default to 1. Also, add an integration test to prove this
scenario and update the JSON schema to verify that the number is a valid
integer.

FIX: THREESCALE-3382

Signed-off-by: Eloy Coto <eloy.coto@gmail.com>