Return 404 to client if no upstream defined.
Due to the large use of routing policy, nowadays and upstream maybe is
not defined at all, APIcast returns 200 back and logs a new message
that a upstream is not defined.

Due to the work in THREESCALE-3775, the team decided to return a 404
back to the client to be aware that the upstream is not correct. Is not
a 502 status code due to the gateway is not defined at all and can be
confusing.

Related-to THREESCALE-3775
Related-to https://github.com/3scale/porta/pull/1385

Signed-off-by: Eloy Coto <eloy.coto@gmail.com>