Usage: Add two new methods
usage:encode_format: To return an string with the format encoded
url-encoded. It's used to communicate with backend and for caching
pourposes.

usage:get_max_delta: This method return the max number of deltas, it's
used to predict the number of hits in a metric in rate_limit_headers
policy.

Signed-off-by: Eloy Coto <eloy.coto@gmail.com>