Resolver: using Fully qualified names.
When adding a dot to the end of the domain, the search string from
resolv.conf should not be added, because the domain is absolute.

Adding this, will save a few DNS request to the DNS server because no
cluster.local, etc.. will be added to the Qname.

More info in relation to this:
https://superuser.com/a/1468139

Fix THREESCALE-4752

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>