APICAST_SERVICES_FILTER_BY_URL fix when remote_v2 is in use
When remote_v2 config is in use, the APICAST_SERVICES_FILTER_BY_URL
parameter was not in use. The main reason was because  the IDs are
filtered out, but the host is not available on api/services.json.

This is a issue when multiple services are in place, and all of them
request info to openid connect, where things are getting super slow, but
in this case, the request to
admin/api/services/${serice id}/proxy/configs/staging/latest.json need to be done
to be able to filter it out.

FIX THREESCALE-6139

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>