usage: fix issues on encoding
Usage keys are based on ngx.encode_args, BUT due to the Luajit update
the data is not always the same, so the following usage metatable:

```
{
  metric_a=1,
  metric_b=1,
}
```

can be `metric_a=1&metric_b=1` or `metric_b=1&metric_a=1`, so the cached
key is not the same, and auth call will happen a few more times when it
shouldn't.

The way that I followed is the same as Openresty, and the escape_uri
should always be a string due to the usage method.

Fix THREESCALE-7122

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>