Openssl: avoid ffi.metatype
For some reason I had issues when using ffi.metatypes to interact with
the ffi interface. With this change,the store will not use ffi.metatype
and it'll send the object natively to the ffi functions.

Other option was to use self.cdata, but it's not clear at all, and this
makes more sense.

Inspired in the work made by https://github.com/fffonion/lua-resty-openssl

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>