[conditional_policy] Do not terminate the chain when failed to load policy
Previously, we used "load_policy" to load and create a new conditional
chain. However, if one or more policies in the chain failed to load,
it would terminate the entire chain.

With this PR, we will use "add_policy" instead, which will skip the failed
policy and load the next policy in the chain.