[http_ng] Fix regression
In PR #1503 we return nil and error in case of an error, however some
code paths do not handle the error but instead check the error field of
the response object and lead to unexpected behavior.

With this PR, we will return a response object with the error field set.