[policy_chain] Log policy chain error
In PR #1286, we wrapped the policy chain call in a pcall but did not handle
the returned error and ignored any exceptions or errors that arose during
execution. This also means that the code that should have failed and terminated
the request is now ignored and allows the chain to move on to the next policy
and allow the request to reach upstream unless on_failed policy is in
the chain.

The design decision is questionable, but to maintain current behavior we will
log errors and ensure that any critical code paths terminate the request.